Move utils to tune_utils.
